SCRUM-97 refactor: Refactor and relocate shared API error types.
Moved `BaseServerError` and `MessageField` types to a new `baseApi.types` file for better modularity and reusability. Updated imports across the codebase to reflect the new structure. Removed redundant type definitions in `authApi.types`.
