+++ src/store/services/api/baseApi/baseApi.types.ts
@@ -1,0 +1,12 @@
+export type BaseServerError = {
+  data: {
+    statusCode: number;
+    messages: MessageField[];
+    error: string;
+  };
+};
+
+export type MessageField = {
+  message: string;
+  field: string;
+};
